[REFACTOR] code in bootstrapy module
