+++ src/flask/app.py
@@ -4,32 +4,18 @@ from flask import Flask, request, render_template, jsonify
 import subprocess
 import re, sys
 from random import random
+from bootstrapy import bootstrapy
 
 app = Flask(__name__)
 
-def build_2x2_grid(c1, c2, c3, c4):
-    return build_container (build_row (build_col_md(c1, 6, 100) + build_col_md(c2, 6, 100), 50) + build_row (build_col_md(c3, 6, 100) + build_col_md(c4, 6, 100), 50))
-
-def build_container(content):
-    return '<div class="container-fluid">' + content + '</div>'
-
-def build_col_md(content, nb, height):
-    return '<div class="col-md-' + str(nb) + '" style="height: ' + str(height) + '%">'+ content + '</div>'
-
-def build_row(content, height):
-    return '<div class="row" style="height: ' + str(height) + '%">'+ content + '</div>'
-
-def build_well(content):
-    return '<div class="well" style="height: 90%">' + content + '</div>'
-
 @app.route('/test')
 def test():
-    well = build_well('<span name="hw"></span>')
-    well_2x2 = build_2x2_grid(well, well, well, well)
+    well = bootstrapy.build_well('<span name="hw">Loading...</span>')
+    well_2x2 = bootstrapy.build_2x2_grid(well, well, well, well)
 
     hello_world_module = {'interval': '1000', 'function': 'hello_func', 'url': 'hello-world', 'id': 'hw'}
 
-    return render_template("main.html", content=well_2x2, modules=[ping_module, hello_world_module])
+    return render_template("main.html", content=well_2x2, modules=[hello_world_module])
 
 @app.route('/get_from_module/<module>')
 def get_from_module(module):
@@ -39,20 +25,6 @@ def get_from_module(module):
 @app.route('/')
 def main():
     return render_template("main.html", content=well_4x4, modules=[ping_module, random_module])
-
-@app.route('/ping')
-def ping():
-    try:
-        out = subprocess.check_output(["ping", "-c", "1", 'google.fr']).decode('utf-8')
-        regex = re.compile(r'time=([0-9]*).')
-        res = 0
-        count = 0
-        for m in re.findall (regex, out):
-            count = count + 1
-            res += int(m)
-        return out + '<br/>moy = '+ str(res / count) + 'ms'
-    except Exception as e:
-        return "Unexpected error:" + str(e)
 
 if __name__ == '__main__':
     app.run(host='0.0.0.0', port=80, debug=True, threaded=True)
